restructure proj
