+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cracky",
+  "name": "buckmed",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
