--- src/components/layouts/PlatformBase.js
@@ -1,13 +1,0 @@
-import React, { Fragment } from 'react';
-import NavBar from './NavBar';
-import Footer from './Footer';
-
-export default ({ children }) => {
-  return (
-    <Fragment>
-      <NavBar />
-      {children}
-      <Footer />
-    </Fragment>
-  );
-};
