--- src/components/pages/AdminBoard/Pages/Appointment.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import AdminBase from '../Base';
-
-const Appointment = () => {
-  return <AdminBase>Appointment page</AdminBase>;
-};
-
-export default Appointment;
