--- src/components/pages/AdminBoard/Pages/Doctors.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import AdminBase from '../Base';
-
-const Doctors = () => {
-  return <AdminBase>Doctors page</AdminBase>;
-};
-
-export default Doctors;
