--- src/components/pages/AdminBoard/Pages/Finance.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import AdminBase from '../Base';
-
-const Finance = () => {
-  return <AdminBase>Finance Page</AdminBase>;
-};
-
-export default Finance;
