--- src/components/pages/AdminBoard/Pages/History.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import AdminBase from '../Base';
-
-const History = () => {
-  return <AdminBase>History Page</AdminBase>;
-};
-
-export default History;
