--- src/components/pages/AdminBoard/Pages/Patients.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import AdminBase from '../Base';
-
-const Patients = () => {
-  return <AdminBase>Patients page</AdminBase>;
-};
-
-export default Patients;
