--- src/components/pages/AdminBoard/index.js
@@ -1,8 +1,0 @@
-import Dashboard from './Pages/Dashboard';
-import Appointment from './Pages/Appointment';
-import Doctors from './Pages/Doctors';
-import Patients from './Pages/Patients';
-import Finance from './Pages/Finance';
-import History from './Pages/History';
-
-export default { Dashboard, Appointment, Doctors, Patients, Finance, History };
