--- src/components/pages/auth/index.js
@@ -1,7 +1,0 @@
-import Login from './Login'
-import Register from './Register'
-
-export default {
-    Login,
-    Register
-}
