--- src/components/pages/index.js
@@ -1,6 +1,0 @@
-import LandingPage from './PlatformView';
-import AuthPage from './auth';
-import Contact from './PlatformView/Contact';
-import AdminBoard from './AdminBoard';
-
-export { LandingPage, AuthPage, Contact, AdminBoard };
