+++ src/pages/doctorBoard/Appointment.js/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Appointment = () => {
+  return <div>Appointment page</div>;
+};
+
+export default Appointment;
