+++ src/pages/doctorBoard/Dashboard/Welcome.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const Welcome = () => {
+  return (
+    <div className='container-fluid admin--welcome-message'>
+    <div className="row w-100 h-100 px-3 m-auto align-items-center">
+      <div className='col-12 col-md-6'>
+        <h1>Hello Shaun!</h1>
+        <p>
+          Here are your important tasks, updates and alerts. <br />   You can
+          set your in app preferences here
+        </p>
+      </div>
+      </div>
+    </div>
+  );
+};
+
+export default Welcome;
