+++ src/pages/doctorBoard/Finance/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Finance = () => {
+  return <div>Finance Page</div>;
+};
+
+export default Finance;
