+++ src/pages/doctorBoard/History/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const History = () => {
+  return <div>History Page</div>;
+};
+
+export default History;
