+++ src/pages/doctorBoard/Patient/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Patients = () => {
+  return <div>Patients page</div>;
+};
+
+export default Patients;
