+++ src/pages/errors/NotFoundView.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const NotFoundView = (props) => {
+  return <div>Not Found Page</div>;
+};
+
+NotFoundView.propTypes = {};
+
+export default NotFoundView;
